rename fauxhdfbuilder -> builder
